Replaced the old logger.py with the new logger, that is previously named new_logger.py. Added Documentation for logger.py. Clean up print.
